Cleaned Code, centralized code, bug fixing
