+++ expressServer/database/database.js
@@ -8,9 +8,7 @@ let connection;
 const connectDB = async () => {
   try {
     if (!connection) {
-      console.log("Connecting to MongoDB");
       connection = await mongoose.connect(process.env.MONGODB_URI);
-      console.log("Connected to MongoDB");
     }
     return connection;
   } catch (err) {
@@ -23,7 +21,6 @@ const closeDB = async () => {
   try {
     if (connection) {
       await connection.disconnect();
-      console.log("Disconnected from MongoDB");
     }
   } catch (err) {
     console.error('Could not disconnect from MongoDB', err);
