+++ expressServer/index.js
@@ -1,5 +1,4 @@
 const express = require('express');
-const mongoose = require('mongoose');
 const cors = require('cors');
 const dotenv = require('dotenv');
 const {connectDB, closeDB} = require('./libraries/database.js')
@@ -11,13 +10,6 @@ const port = process.env.PORT || 5000;
 
 app.use(cors());
 app.use(express.json());
-
-const userSchema = new mongoose.Schema({
-  username: { type: String, required: true },
-  password: { type: String, required: true },
-});
-
-const User = mongoose.model('User', userSchema);
 
 app.post('/api/login', async (req, res) => {
   const { username, password } = req.body;
@@ -59,6 +51,8 @@ app.get('/api/users', async (req, res) => {
   } catch (error) {
     console.error('Error fetching users:', error);
     res.status(500).json({ message: 'Internal server error' });
+  } finally {
+    closeDB();
   }
 });
