+++ expressServer/libraries/saltUsername.js
@@ -1,0 +1,8 @@
+const crypto = require('crypto');
+
+const saltUser = () => {
+    const salt = crypto.randomBytes(32).toString('hex');
+    return (salt)
+}
+
+module.exports = { saltUser };
